+++ CHANGELOG.md
@@ -1,5 +1,12 @@
 # Changelog
 
+## [Unreleased] - YYYY-MM-DD
+
+### Changed
+
+* The values of `host`, `exchange` and `queue` are now required if the options were provided. Previously it was possible to run the command with for example `-h` and not providing the URL - which led to errors. 
+* Unified the closing behaviour of `produce` and `consume` commands - they report the error in the same way and shut down the process with exit code `1`.
+
 ## [0.2.2] - 2020-06-29
 
 Note: 0.2.1 was wrongly released and is replaced by 0.0.2